src/main/java/pt/ipp/isep/dei/project1/readers/readerjson/RoomFromJsonFile.java
package pt.ipp.isep.dei.project1.readers.readerjson;

public class RoomFromJsonFile {

    private String id;
    private String description;
    private double floor;
    private double width;
    private double length;
    private double height;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public double getFloor() {
        return floor;
    }

    public void setFloor(double floor) {
        this.floor = floor;
    }

    public double getWidth() {
        return width;
    }

    public void setWidth(double width) {
        this.width = width;
    }

    public double getLength() {
        return length;
    }

    public void setLength(double length) {
        this.length = length;
    }

    public double getHeight() {
        return height;
    }

    public void setHeight(double height) {
        this.height = height;
    }
}
